+++ src/main/java/com/mcommings/campaigner/services/WealthService.java
@@ -3,23 +3,33 @@ package com.mcommings.campaigner.services;
 import com.mcommings.campaigner.interfaces.IWealth;
 import com.mcommings.campaigner.models.RepositoryHelper;
 import com.mcommings.campaigner.models.Wealth;
+import com.mcommings.campaigner.models.repositories.ICityRepository;
 import com.mcommings.campaigner.models.repositories.IWealthRepository;
 import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Service;
 
+import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.List;
 
-import static com.mcommings.campaigner.ErrorMessage.*;
+import static com.mcommings.campaigner.enums.ErrorMessage.*;
+import static com.mcommings.campaigner.enums.ForeignKey.FK_WEALTH;
 
 @Service
 public class WealthService implements IWealth {
 
     private final IWealthRepository wealthRepository;
+    private final ICityRepository cityRepository;
 
     @Autowired
-    public WealthService(IWealthRepository wealthRepository) {this.wealthRepository = wealthRepository;}
+    public WealthService(IWealthRepository wealthRepository, ICityRepository cityRepository) {
+        this.wealthRepository = wealthRepository;
+        this.cityRepository = cityRepository;
+    }
+
     @Override
     public List<Wealth> getWealth() {
         return wealthRepository.findAll();
@@ -28,10 +38,10 @@ public class WealthService implements IWealth {
     @Override
     @Transactional
     public void saveWealth(Wealth wealth) throws IllegalArgumentException, DataIntegrityViolationException {
-        if(RepositoryHelper.nameIsNullOrEmpty(wealth)) {
+        if (RepositoryHelper.nameIsNullOrEmpty(wealth)) {
             throw new IllegalArgumentException(NULL_OR_EMPTY.message);
         }
-        if(RepositoryHelper.nameAlreadyExists(wealthRepository, wealth)) {
+        if (RepositoryHelper.nameAlreadyExists(wealthRepository, wealth)) {
             throw new DataIntegrityViolationException(NAME_EXISTS.message);
         }
 
@@ -41,10 +51,12 @@ public class WealthService implements IWealth {
     @Override
     @Transactional
     public void deleteWealth(int wealthId) throws IllegalArgumentException, DataIntegrityViolationException {
-        if(RepositoryHelper.cannotFindId(wealthRepository, wealthId)) {
+        if (RepositoryHelper.cannotFindId(wealthRepository, wealthId)) {
             throw new IllegalArgumentException(DELETE_NOT_FOUND.message);
         }
-        //TODO: check if foreign key
+        if (RepositoryHelper.isForeignKey(getReposWhereWealthIsAForeignKey(), FK_WEALTH.columnName, wealthId)) {
+            throw new DataIntegrityViolationException(DELETE_FOREIGN_KEY.message);
+        }
 
         wealthRepository.deleteById(wealthId);
     }
@@ -52,11 +64,16 @@ public class WealthService implements IWealth {
     @Override
     @Transactional
     public void updateWealth(int wealthId, Wealth wealth) throws IllegalArgumentException, DataIntegrityViolationException {
-        if(RepositoryHelper.cannotFindId(wealthRepository, wealthId)) {
+        if (RepositoryHelper.cannotFindId(wealthRepository, wealthId)) {
             throw new IllegalArgumentException(UPDATE_NOT_FOUND.message);
         }
 
         Wealth wealthToUpdate = RepositoryHelper.getById(wealthRepository, wealthId);
         wealthToUpdate.setName(wealth.getName());
+    }
+
+    private List<CrudRepository> getReposWhereWealthIsAForeignKey() {
+        List<CrudRepository> repositories = new ArrayList<>(Arrays.asList(cityRepository));
+        return repositories;
     }
 }
